feat: add info tx types
